Rename Main container to App in CM
