+++ packages/core/content-manager/admin/src/pages/App/actions.js
