+++ packages/core/content-manager/admin/src/pages/App/constants.js
@@ -1,0 +1,3 @@
+export const GET_DATA = 'ContentManager/App/GET_DATA';
+export const GET_DATA_SUCCEEDED = 'ContentManager/App/GET_DATA_SUCCEEDED';
+export const RESET_PROPS = 'ContentManager/App/RESET_PROPS';
