+++ packages/core/content-manager/admin/src/pages/App/reducer.js
