--- packages/core/content-manager/admin/src/pages/Main/constants.js
@@ -1,3 +1,0 @@
-export const GET_DATA = 'ContentManager/Main/GET_DATA';
-export const GET_DATA_SUCCEEDED = 'ContentManager/Main/GET_DATA_SUCCEEDED';
-export const RESET_PROPS = 'ContentManager/Main/RESET_PROPS';
