--- packages/core/content-manager/admin/src/pages/Main/selectors.js
@@ -1,20 +1,0 @@
-import { createSelector } from 'reselect';
-import pluginId from '../../pluginId';
-import { initialState } from './reducer';
-
-const selectMainDomain = () => state => {
-  return state[`${pluginId}_main`] || initialState;
-};
-
-const makeSelectMain = () =>
-  createSelector(selectMainDomain(), substate => {
-    return substate;
-  });
-
-const makeSelectModelAndComponentSchemas = () =>
-  createSelector(selectMainDomain(), ({ components, models }) => ({
-    schemas: [...components, ...models],
-  }));
-
-export default makeSelectMain;
-export { makeSelectModelAndComponentSchemas, selectMainDomain };
