+++ packages/core/content-manager/admin/src/reducers.js
@@ -1,4 +1,4 @@
-import mainReducer from './pages/Main/reducer';
+import appReducer from './pages/App/reducer';
 import editViewLayoutManagerReducer from './pages/EditViewLayoutManager/reducer';
 import listViewReducer from './pages/ListView/reducer';
 import rbacManagerReducer from './hooks/useSyncRbac/reducer';
@@ -6,7 +6,7 @@ import editViewCrudReducer from './sharedReducers/crudReducer/reducer';
 import pluginId from './pluginId';
 
 const reducers = {
-  [`${pluginId}_main`]: mainReducer,
+  [`${pluginId}_app`]: appReducer,
   [`${pluginId}_listView`]: listViewReducer,
   [`${pluginId}_rbacManager`]: rbacManagerReducer,
   [`${pluginId}_editViewLayoutManager`]: editViewLayoutManagerReducer,
